@@ -13,11 +13,11 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=8.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "5.7",
-        "squizlabs/php_codesniffer": "2.8.0"
+        "phpunit/phpunit": "^9.0",
+        "squizlabs/php_codesniffer": "^3.6"
     },
     "autoload": {
         "psr-4": {
